requirements.txt
numpy~=1.26.4
pandas~=2.2.2
matplotlib~=3.9.0
pillow~=10.3.0
bcrypt~=4.1.3
tensorflow~=2.17.0
pyserial~=3.5
psutil~=6.0.0
ttkthemes~=3.2.2
aiofiles~=24.1.0
aiocsv~=1.3.2